Desktop Application Cleaning
To make development easy until a version with reduced dependencies is
sorted out, I included a copy of openFrameworks 0.7.4 for vs2010 which
includes the addons needed for the desktop app. To use the current
version of the app, use the exe in the bin dir of openRowland, or unzip
openFrameworks, and paste the openRowland directory to
openFrameworks\apps\myApps. the app is not included in the 7zip file
because I will update the binaries as I go along so there will always a
build for demonstration purposes. All right, wall of text is finished.