Added hardware block diagram, and made the CCD demo work.
Added sketch of rough hardware architecture, as well as updated CCD test
code to correctly deal with the ROG signal line. CCD test code now is
functional, and demonstrates relative timing of different signals, as
well as importance of exact duration of the ROG signal (and
insignificance of the clock speed...).